Address PR feedback
- Move the new test class to a separate file
