+++ src/System.Collections/tests/Generic/SortedSet/SortedSet.Generic.Tests.cs
@@ -283,31 +283,4 @@ namespace System.Collections.Tests
 
         #endregion
     }
-
-    public class SortedSet_ICollection_NonGeneric_Tests : ICollection_NonGeneric_Tests
-    {
-        protected override bool ICollection_NonGeneric_CopyTo_ArrayOfEnumType_ThrowsArgumentException { get { return true; } }
-        protected override bool Enumerator_Current_UndefinedOperation_Throws { get { return true; } }
-
-        protected override void AddToCollection(ICollection collection, int numberOfItemsToAdd)
-        {
-            int seed = numberOfItemsToAdd * 34;
-            for (int i = 0; i < numberOfItemsToAdd; i++)
-                ((SortedSet<string>)collection).Add(CreateT(seed++));
-        }
-
-        protected override ICollection NonGenericICollectionFactory()
-        {
-            return new SortedSet<string>();
-        }
-
-        protected string CreateT(int seed)
-        {
-            int stringLength = seed % 10 + 5;
-            Random rand = new Random(seed);
-            byte[] bytes = new byte[stringLength];
-            rand.NextBytes(bytes);
-            return Convert.ToBase64String(bytes);
-        }
-    }
 }
